pyproject.toml
[project]
name = "ecommerce-data-pipeline"
version = "0.1.0"
description = "Ecommerce data pipeline with Great Expectations validation and FastAPI serving layer."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.7",
    "pandas>=2.2",
    "httpx>=0.27",
    "tenacity>=9.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "respx>=0.21",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
